Treat HiDPI and monitor settings better.
Some sizes were x2 to match HiDPI, this is no longer the case, from now
on an X resource should indicate the DPI and the program should know to
handle the size (and display x2 if necessary).

Other then that, the specific monitors configuration is now not a part
of the repository but rather in a $HOME/.monitors script which should
configure the monitors and then return their names. This allows the same
dotfiles for multiple machines with different monitor configurations.