sway: Add sway, including waybar and swaylock
This is now intended to be used instead of i3, which is kept in the repo
but might be removed in the future.

The sway configuration is the same as the i3 configuration, with a few
additions:
- Use a special media workspace
- Add keybindings to move/rename entire workspaces
- Add a keybinding to launch a floating shell
- Add keybindings to control the media being played
- A keybinding to keep a floating window stickied
- Minor appearance tweaks

Also given that sway is in charge of configuring more stuff than i3,
this also adds configuration for:
- Automatically going into a swaylock lockscreen after inactivity
- Configure the laptop display to close/open based on its lid
- Language configuration
- Input configuration

This commit also adds waybar configuration, which is a change from i3's
polybar. This also includes a new script to check the internet
connectivity and an improvement to the Arch package update checking
script.