neovim: Switch from fzf.vim to fzf-lua
Basically a lua plugin that does kind of the same thing, but in lua and
not in vimscript.

Changes/improvements include:
- LSP fzf functions
- Fuzzy search current selection
- Same style in the preview buffer
- Switch to rg from ag for searches

Note that I still didn't re-implement the :FromIndex and :AgIndex
command from vimscript, so their previous implementations are included
in a comment. They should be re-implemented in the future (once the need
arises).