neovim: Automatically enforce character limit for comments and text
When editing comments and text(/markdown) files, automatically go down a
line when surprassing the character limit per line (80).

This is done because enforcing it for code is too complex to be done
automatically, but not for text.

Also, when entering a new line to enforce character limit, try to
align to lists.