neovim: Switch from packer to lazy plugin manager
Not sure if the value from switching is that big, but mainly so far:

- Creates a lock file that denotes the commit for each plugin. Useful if
  I start using neovim on a different machine and I don't want things to
  suddenly break because the config is not adapted to a later plugin
  version.
- Works well when running neovim for the first time, no need to do the
  open & reopen dance until all plugins are working like was needed in
  packer.
  For this I also improved a bit the code that handles running on a new
  machine. I even tried to make it work in case I ever decide to make an
  AUR package for lazy.nvim.

Other than this lazy also seems to have more hype I guess.

I will look a bit into the lazy loading in the future, but honestly I
don't feel that I have that many plugins and use cases for that. I guess
maybe for different file types?