sway: waybar: Show date on click instead of full time
I feel that I need the date much more than I need the full time, and
hovering and waiting for the tooltip to show is slower than clicking.

I wish there was a way to configure waybar to show the tooltip on
click/change hover timeout to 0, but this will do for now. (I guess if I
want I could try to add this feature to waybar)