neovim: Use ctags for fzf tags bindings if LSP is not available
This is not fully robust as I guess it could be possible that the same
session has both a ctags and an LSP files open, but for now this is good
enough.

Also unfortunately it seems that fzf-lua handles ctags a bit ugly and
that it varies according to the format of the ctags file, but I'm not
sure there's much to do about that.