sway: Add script to manage laptop display
The script manages laptop display to turn it off if the lid is closed or if
other displays are connected, but otherwise on. Can be toggled to still
have laptop display on even if other displays are connected using the
`laptop_display_toggle` fish function.

This replaces the previous config of just setting laptop display status
based on whether lid is open or closed.

This should make it nicer to work with laptop connected to an external
display but with its lid open, e.g. will give access to fingerprint
sensor.